+++ src/commands/extractTranslationString/utils/stringUtils.ts
@@ -22,7 +22,15 @@ export const stripFirstLast = (s: string): string =>
 export const stripQuotes = (s: string): string =>
   isSingleQuoted(s) || isDoubleQuoted(s) ? stripFirstLast(s) : s;
 
-export const wrapWithTranslationHook = (s: string) => `t(${s})`;
+export const wrapWithTranslationHook = (s: string, args?: string[]) => {
+  if (args && args.length > 0) {
+    const argumentsObject = `{${args.join(", ")}}`;
+
+    return `t(${removeCurlyBracketsFromString(s)}, ${argumentsObject})`;
+  }
+
+  return `t(${s})`;
+};
 
 export const wrapWithCurlyBrackets = (s: string) => `{${s}}`;
 
@@ -32,3 +40,26 @@ export const consolidateMultiLineString = (s: string): string => {
 
   return trimmedLines.join(" ");
 };
+
+export const getArgumentsFromJsxStringLiteral = (s: string): string[] => {
+  let insideCurlyBrackets = false;
+  let currentArgument = "";
+  let args: string[] = [];
+
+  for (const character of s) {
+    if (character === leftCurlyBracket) {
+      insideCurlyBrackets = true;
+    } else if (character === rightCurlyBracket) {
+      insideCurlyBrackets = false;
+      args.push(currentArgument);
+      currentArgument = "";
+    } else if (insideCurlyBrackets) {
+      currentArgument += character;
+    }
+  }
+
+  return args;
+};
+
+export const removeCurlyBracketsFromString = (s: string): string =>
+  s.replace(/\{/g, "").replace(/\}/g, "");